intiatl data for users and posts working
